Fixed a little issue with the dBFS calculation due to implicit conversions from uint to int and vice versa, all working now

+++ include/peak_detect.h
@@ -21,13 +21,13 @@
 /* FUNCTION DECLARATIONS */
 
 // Find the Peak in a buffer, Find the decay of the last sample, return the larger
-uint32_t peak_output(volatile int32_t* buffer, uint8_t buffer_size, uint8_t decay_shift);
+uint16_t peak_output(volatile int16_t* buffer, uint8_t buffer_size, uint8_t decay_shift);
 
 // Take a ADC Reading and Convert to 16 bit scale dBFS
-int32_t dbfs_output(uint32_t input);
+int16_t dbfs_output(uint16_t input);
 
 // Print a graphic line proportional to the input
-void pretty_print(int32_t sample, uint8_t scale_shift);
+void pretty_print(uint16_t sample, uint8_t scale_shift);
 
 
 #endif /* PEAK_DETECT_H_ */
